Dropdown format bar for h6 tag match redactor field
Craft defines CSS for the `h6` tag in the redactor field to text-transform: none. It doesn't make this same change to the `h6` selector in the format bar. This makes the `h6` in the dropdown bar UPPERCASE, inherited from the Redactor default.

The `h6` should be the same visually in the dropdown bar and the text entry area for Redactor in the control panel.